Added more risk outputs to the extract API [skip hazardlib]
